Add Users API
Routes map URLs to controller functions and are defined in routes/api.php.
Requests define a set of validators to validate any request to a Route.
Resources define an API response. They map from database models to JSON
objects.
Controllers define a set of methods to be associated with particular
routes. They define the logic that is run when a route receives a valid
request.

This commit adds Routes, Requests, a Resource, and a Controller so that
students can be added, modified, and searched for.

The code has not yet been tested (it's late and I don't feel like
figuring out the testing framework at this moment) so don't expect it to
work especially well.