Differentiate between Students and Users
Students are considered only as the subject of events. They are not for
authentication--students cannot log in. Only users can log in.

Users are either mentors or tools. They have permission to check
students in or out, add students, or edit students. Users are the
registrars in events -- they are responsible for registering events, but
they themselves are not the subject of those events.