[Web] Improve progress bar on import-students
Previously, the user was shown an indeterminate spinner while the
imported students were uploaded. This was better than nothing, since it
gave the user a sense that something was going on, and the app hadn't
just frozen.

However, it would be better to show the user exactly how much progress
has been made, and how much is left to be done. To that end, this patch
replaces the indeterminate progress-spinner with a determinate
progress-bar.