[Api] Fix database timezone
All timestamps were being saved in the database 7 hours in the future.
The issue was a double-conversion to/from UTC time.

As best I can tell, the faulty behavior was as follows:
1. The client sends the server a unix timestamp.
2. The server formats it into a UTC datetime string and sends it to the
    database.
3. The database assumes the given datetime is in local time, and
   converts it to UTC

The solution is to configure the database such that it recognizes that
datetime strings from the server are already in UTC.

(I hate timezones.)