[Web] Increase polling frequency
At a recent meeting, there were incidences of simultanous checkouts
within the poll interval--i.e. mentor A checked out a student right
after mentor B checked out the same student, before mentor A's device
had polled for the updated information. This causes desync, and a 422
http error. The temporary solution is to poll more frequently, so that
the issue is less likely.

We should also probably handle the 422 errors gracefully, but if I
decide to do that, that'll happen in a separate commit.