[Web] Lock interface after polling update
Carl was noticing that an update from another mentor checking a student
in or out could cause the interface to shift right as the user was about
to push a button, potentially causing the wrong student to be checked in
or out.

The solution is to disable the interface for 100ms after such an update
is received, to give the user time to readjust.