[Web] Gracefully handle simultaneous attendance events
If the server flags that a duplicate attendance event already exists,
instead of erroring out the client, we just refresh the student's status
in the attendance list. To the user, it appears that the event was
registered, and the only difference from normal operation is that the
undo does not appear (since we are not actually registering a new event,
we're just updating the page status to account for previously entered
events).

Fixes #75