[Api] Remove authentication from server info api endpoint
We were getting an error on the main login page because we received an
HTTP 401 (Unauthenticated) response from the info endpoint.
I don't see any reason to hide the server version behind authentication,
so instead of updating the client to handle the error, I'll just remove
the authentication from the endpoint.