[Meta] Force cache revalidation
Use the Cache-Control header to force the user's browser to revalidate
the client source files against the server. This way, whenever I release
a new version of the attendance manager, the browser will automatically
download the update instead of using the old version from its cache.
This shouldn't have much bandwidth overhead since the browser just needs
to send the ETag of its cache to the server, so it's only redownloading
the whole website when the files have changed.