[Api] Fix laravel tests randomly failing
Students created by the student factory are randomly soft-deleted. This
is expected, as the point of the student factory is to populate the
database with generated testing data that matches the real world as
close as possible, and in the real world some students are soft-deleted.

However, the test_destroy test of the StudentControllerTest did not
ensure that the student it was trying to delete was not already deleted.
So whenever the student was already deleted, the api returned a 404 and
the test failed.

The solution is to just ensure the student factory doesn't soft-delete
any students when running this test. This patch also adds some
additional checks to ensure the student isn't already soft-deleted, and
to ensure that we do get a 404 if we try to soft-delete twice.