[Meta] Force cache validation on all files
Originally, we only set the must-revalidate header on .html, .css, and
.js files. However, most of our endpoints don't have a .html extension
(for example /attendance/check-in), so they weren't getting the
must-revalidate header. This would cause errors whenever an update was
deployed, since the browser would use the old, cached html data which
referred to a .js file that no longer existed on the server. The browser
would try to revalidate the .js file, get back the 404 html response,
and error out.