[Api] Use single query for student last_check_in/out
Previously, when asked to return a list of all students, we would
perform a database query per-student to retrieve the last check in and
last check out, which we include in the student response. This is very
inefficient, and as the number of registered students increases, this
problem becomes worse and worse.

The solution is to just retrieve all of the most recent attendance
events for each student in just one query, then manually append these
values to the student models.

Fixes #81